// packages/layerzero-definitions/src/enums/chain.ts
export enum Chain {
    ETHEREUM = 'ethereum',
    BSC = 'bsc',
    AVALANCHE = 'avalanche',
    POLYGON = 'polygon',
    ZKPOLYGON = 'zkpolygon',
    ARBITRUM = 'arbitrum',
    OPTIMISM = 'optimism',
    FANTOM = 'fantom',
    SWIMMER = 'swimmer',
    DFK = 'dfk',
    HARMONY = 'harmony',
    ARCANA = 'arcana',
    DEXALOT = 'dexalot',
    CASTLECRUSH = 'castlecrush',
    GNOSIS = 'gnosis',
    CELO = 'celo',
    MOONBEAM = 'moonbeam',
    BOBA = 'boba',
    PORTAL = 'portal',
    AURORA = 'aurora',
    INTAIN = 'intain',
    FUSE = 'fuse',
    ASTAR = 'astar',
    ZKSYNC = 'zksync',
    SHRAPNEL = 'shrapnel',
    DOS = 'dos',
    APTOS = 'aptos',
    KLAYTN = 'klaytn',
    METIS = 'metis',
    COREDAO = 'coredao',
    OKX = 'okx',
    METER = 'meter',
    GOERLI = 'goerli',
    SOLANA = 'solana',
    BASE = 'base',
    ZKCONSENSYS = 'zkconsensys',
    SEPOLIA = 'sepolia',
    MOONRIVER = 'moonriver',
    AAVEGOTCHI = 'aavegotchi',
    SCROLL = 'scroll',
    CATHAY = 'cathay',
    KAVA = 'kava',
    TENET = 'tenet',
    CANTO = 'canto',
    ORDERLY = 'orderly',
    NOVA = 'nova',
    BLOCKGEN = 'blockgen',
    MERITCIRCLE = 'meritcircle',
    MANTLE = 'mantle',
    HUBBLE = 'hubble',
    ZORA = 'zora',
}