// packages/layerzero-definitions/src/enums/chain-key.ts
// APPEND ONLY LIST
// verify twice that the value is correct
export enum ChainKey {
    ETHEREUM = 'ethereum',
    GOERLI = 'goerli',
    SEPOLIA = 'sepolia',
    BSC = 'bsc',
    BSC_TESTNET = 'bsc-testnet',
    AVALANCHE = 'avalanche',
    FUJI = 'fuji',
    POLYGON = 'polygon',
    MUMBAI = 'mumbai',
    ARBITRUM = 'arbitrum',
    ARBITURM_GOERLI = 'arbitrum-goerli',
    OPTIMISM = 'optimism',
    OPTIMISM_GOERLI = 'optimism-goerli',
    FANTOM = 'fantom',
    FANTOM_TESTNET = 'fantom-testnet',
    SWIMMER = 'swimmer',
    SWIMMER_TESTNET = 'swimmer-testnet',
    DFK = 'dfk',
    DFK_TESTNET = 'dfk-testnet',
    HARMONY = 'harmony',
    HARMONY_TESTNET = 'harmony-testnet',
    DEXALOT = 'dexalot',
    DEVALOT_TESTNET = 'dexalot-testnet',
    CELO = 'celo',
    ALFAJORES = 'alfajores',
    MOONBEAM = 'moonbeam',
    MOONBASE = 'moonbase',
    MOONRIVER = 'moonriver',
    INTAIN = 'intain',
    INTAIN_TESTNET = 'intain-testnet',
    FUSE = 'fuse',
    FUSESPARK = 'fusespark',
    GNOSIS = 'gnosis',
    CHIADO = 'chiado',
    APTOS = 'aptos',
    APTOS_TESTNET = 'aptos-testnet',
    KLAYTN = 'klaytn',
    KLAYTN_BAOBAB = 'klaytn-baobab',
    METIS = 'metis',
    METIS_GOERLI = 'metis-goerli',
    COREDAO = 'coredao',
    COREDAO_TESTNET = 'coredao-testnet',
    METER = 'meter',
    METER_TESTNET = 'meter-testnet',
    OKX = 'okx',
    OKX_TESTNET = 'okx-testnet',
    ZKSYNC = 'zksync',
    ZKSYNC_TESTNET = 'zksync-testnet',
    SHRAPNEL = 'shrapnel',
    SHRAPNEL_TESTNET = 'shrapnel-testnet',
    DOS = 'dos',
    DOS_TESTNET = 'dos-testnet',
    ZKEVM = 'zkevm',
    ZKEVM_TESTNET = 'zkevm-testnet',
    NOVA = 'nova',
    TENET = 'tenet',
    TENET_TESTNET = 'tenet-testnet',
    CANTO = 'canto',
    CANTO_TESTNET = 'canto-testnet',
    KAVA = 'kava',
    KAVA_TESTNET = 'kava-testnet',
    SOLANA = 'solana',
    SOLANA_TESTNET = 'solana-testnet',
    LINEA_GOERLI = 'linea-goerli',
    BASE_GOERLI = 'base-goerli',
    AAVEGOTCHI_TESTNET = 'aavegotchi-testnet',
    SCROLL_TESTNET = 'scroll-testnet',
    CATHAY_TESTNET = 'cathay-testnet',
    ORDERLY_TESTNET = 'orderly-testnet',
    BLOCKGEN_TESTNET = 'blockgen-testnet',
    MERITCIRCLE_TESTNET = 'meritcircle-testnet',
    MANTLE = 'mantle',
    MANTLE_TESTNET = 'mantle-testnet',
    HUBBLE = 'hubble',
    HUBBLE_TESTNET = 'hubble-testnet',
    ZKCONSENSYS = 'zkconsensys',
    BASE = 'base',
    ZORA_TESTNET = 'zora-testnet',
    ZORA = 'zora',
}